pyproject.toml
[tool.poetry]
name = "minimum-wageml"
version = "0.1.0"
description = "use machine learning model to predict who is a minimum wage worker"
authors = ["githubjacky"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
hydra-core = "^1.1.1"
loguru = "^0.7.0"
numpy = "^1.26.0"
pandas = "2.0.0"


[tool.poetry.group.model.dependencies]
scikit-learn = "^1.3.0"
mlflow = "^2.7.1"
xgboost = "^2.0.0"
optuna = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
